+++ tests/com/splunk/HttpServiceTest.java
@@ -20,15 +20,10 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
+import javax.net.ssl.*;
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.PrintStream;
-import java.io.UnsupportedEncodingException;
-import java.util.HashMap;
+import java.security.SecureRandom;
+import java.security.cert.X509Certificate;
 
 public class HttpServiceTest extends SDKTestCase {
     private HttpService httpService;
@@ -85,5 +80,128 @@ public class HttpServiceTest extends SDKTestCase {
         ResponseMessage response = new ResponseMessage(200);
         Assert.assertEquals(response.getStatus(), 200);
         Assert.assertTrue(response.getHeader() != null);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testSSLSocketFactorySetNull(){
+        HttpService.setSSLSocketFactory(null);
+    }
+
+
+
+    @Test
+    public void testSSLSocketFactory() {
+        try {
+            SSLSocketFactory factory = Service.getSSLSocketFactory();
+            SSLSocket socket = (SSLSocket) factory.createSocket("localhost", 8089);
+            String[] protocols = socket.getEnabledProtocols();
+            Assert.assertTrue(protocols.length > 0);
+        }
+        catch (Exception e) {
+            Assert.assertNull(e);
+        }
+    }
+
+    public void validateSSLSocketFactory(SSLSocketFactory factory) {
+        // Backup the old value
+        SSLSocketFactory old = Service.getSSLSocketFactory();
+
+        Service.setSSLSocketFactory(factory);
+        Service s = new Service(service.getHost());
+        s.login(service.getUsername(), service.getPassword());
+        Assert.assertEquals(service.getUsername(), s.getUsername());
+        Assert.assertEquals(service.getPassword(), s.getPassword());
+        Assert.assertEquals(service.getInfo().keySet(), s.getInfo().keySet());
+        Assert.assertEquals(service.getInfo().getVersion(), s.getInfo().getVersion());
+
+        // Restore the old value
+        Service.setSSLSocketFactory(old);
+    }
+
+    @Test
+    public void testCustomSSLSocketFactories() {
+        try {
+            SSLContext sslContext = SSLContext.getInstance("TLS");
+            TrustManager[] byPassTrustManagers = new TrustManager[] {
+                new X509TrustManager() {
+                    public X509Certificate[] getAcceptedIssuers() {
+                        return new X509Certificate[0];
+                    }
+                    public void checkClientTrusted(X509Certificate[] chain, String authType) {}
+                    public void checkServerTrusted(X509Certificate[] chain, String authType) {}
+                }
+            };
+            sslContext.init(null, byPassTrustManagers, new SecureRandom());
+            SSLSocketFactory TLSOnlySSLFactory = sslContext.getSocketFactory();
+            Service.setSSLSocketFactory(TLSOnlySSLFactory);
+
+            validateSSLSocketFactory(TLSOnlySSLFactory);
+        }
+        catch (Exception e) {
+            Assert.assertNull(e);
+        }
+
+        try {
+            SSLContext sslContext = SSLContext.getInstance("SSL");
+            TrustManager[] byPassTrustManagers = new TrustManager[] {
+                new X509TrustManager() {
+                    public X509Certificate[] getAcceptedIssuers() {
+                        return new X509Certificate[0];
+                    }
+                    public void checkClientTrusted(X509Certificate[] chain, String authType) {}
+                    public void checkServerTrusted(X509Certificate[] chain, String authType) {}
+                }
+            };
+            sslContext.init(null, byPassTrustManagers, new SecureRandom());
+            SSLSocketFactory SSLOnlySSLFactory = sslContext.getSocketFactory();
+            Service.setSSLSocketFactory(SSLOnlySSLFactory);
+
+            validateSSLSocketFactory(SSLOnlySSLFactory);
+        }
+        catch (Exception e) {
+            // Swallow exceptions for Java 8, since we know SSLv3 is disabled
+            // by default
+            if (getJavaVersion() < 8) {
+                Assert.assertNull(e);
+            }
+        }
+    }
+
+    public void validateSSLProtocol(Service s, SSLSecurityProtocol securityProtocol) {
+        // Backup the old value
+        SSLSecurityProtocol old = Service.getSslSecurityProtocol();
+
+        Service.setSslSecurityProtocol(securityProtocol);
+
+        s.login(service.getUsername(), service.getPassword());
+        Assert.assertEquals(service.getUsername(), s.getUsername());
+        Assert.assertEquals(service.getPassword(), s.getPassword());
+        Assert.assertEquals(service.getInfo().keySet(), s.getInfo().keySet());
+        Assert.assertEquals(service.getInfo().getVersion(), s.getInfo().getVersion());
+
+        // Restore the value
+        Service.setSslSecurityProtocol(old);
+    }
+
+    @Test
+    public void testSSLSecurityProtocols() {
+        Service s = new Service(service.getHost());
+
+        Integer javaVersion = getJavaVersion();
+        Assert.assertNotNull(javaVersion);
+
+        // TLSv1.1 and TLSv1.2 were added in Java 7
+        if (javaVersion >= 7) {
+            validateSSLProtocol(s, SSLSecurityProtocol.TLSv1_2);
+            validateSSLProtocol(s, SSLSecurityProtocol.TLSv1_1);
+        }
+
+        // TLSv1 is supported in Java 6-8, always check
+        validateSSLProtocol(s, SSLSecurityProtocol.TLSv1);
+
+        // SSLv3 is disabled by default in Java 8
+        if (javaVersion < 8) {
+            validateSSLProtocol(s, SSLSecurityProtocol.SSLv3);
+        }
     }
 }